// oauth-mcp/src/interfaces/auth.ts
export interface PingOneConfig {
  baseUrl: string;
  environmentId?: string;
  clientId: string;
  clientSecret: string;
  tokenIntrospectionEndpoint: string;
  authorizationEndpoint: string;
  tokenEndpoint: string;
}